pyproject.toml
[project]
name = "agents_tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.13"

dependencies = [
    "fastapi>=0.128.0",
    "langchain>=1.2.7,<2.0.0",
    "langchain-openai>=1.1.7",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.21",
    "uvicorn>=0.40.0",
]
